roles/licenses/test-requirements.txt
ansible-lint
docker
flake8
molecule
molecule-docker
netaddr
pytest
pytest-testinfra
tox
yamllint
